requirements.txt
cuda=11.7.1
datasets=2.6.1
evaluate=0.4.0
guildai=0.8.2
numpy=1.21.5
pandas=1.4.4
pyodbc=4.0.34
python=3.9.13
pytorch=1.13.0
transformers=4.24.0